Boss Spawns In Arrows
The "Lasers" use the double jump icon to show direction and moves forward. When they spawn in they aim towards the player and will move in that direction forever.

The plan is to make a deathbox surrounding the boss arena and if they leave the collider they get deleted.

Got a recomendation to make the boss arena another level to have a level 2.